websocket-kafka-rabbitmq-client-python/requirements.txt
eventlet
requests
schedule
websockets
kafka-python